daxctl: Skip over memory failure node status
When trying to match a dax device to a memblock physical address
memblock_in_dev will fail if the the phys_index sysfs file does
not exist in the memblock. Currently the memory failure directory
associated with a node is currently interpreted as a memblock.
Skip over the memory_failure directory within the node directory.

Signed-off-by: Adam Manzanares <a.manzanares@samsung.com>